refactor(ui): extract remote overview tests
